#91 change naming and documentation of diffusivity averaging function
