+++ pybamm/__init__.py
@@ -73,7 +73,6 @@ from .expression_tree.unary_operators import (
     SpatialOperator,
     Gradient,
     Divergence,
-    SpatialAverage,
     grad,
     div,
 )
@@ -108,7 +107,10 @@ from .parameters import standard_parameters
 # Mesh and Discretisation classes
 #
 from .discretisations.base_discretisation import BaseDiscretisation
-from .discretisations.finite_volume_discretisations import FiniteVolumeDiscretisation
+from .discretisations.finite_volume_discretisations import (
+    FiniteVolumeDiscretisation,
+    NodeToEdge,
+)
 from .discretisations.base_mesh import KNOWN_DOMAINS
 from .discretisations.base_mesh import BaseMesh, BaseSubmesh
 from .discretisations.finite_volume_meshes import (
