+++ pybamm/expression_tree/unary_operators.py
@@ -112,43 +112,6 @@ class Divergence(SpatialOperator):
         super().__init__("div", child)
 
 
-class SpatialAverage(SpatialOperator):
-    """A node in the expression tree representing a spatially averaging operator.
-
-    Parameters
-    ----------
-
-    name : str
-        name of the node
-    child : :class:`Symbol`
-        child node
-    averaging_function : method
-        the function used to average; only acts if the child evaluates to a
-        one-dimensional numpy array
-
-    **Extends:** :class:`SpatialOperator`
-    """
-
-    def __init__(self, child, averaging_function):
-        """ See :meth:`pybamm.UnaryOperator.__init__()`. """
-        super().__init__(
-            "spatial average ({})".format(averaging_function.__name__), child
-        )
-        self._averaging_function = averaging_function
-
-    def evaluate(self, t=None, y=None):
-        """ See :meth:`pybamm.Symbol.evaluate()`. """
-        evaluated_child = self.children[0].evaluate(t, y)
-        # If the evaluated child is a numpy array of shape (n,), do the averaging
-        # NOTE: Doing this check every time might be slow?
-        # NOTE: Will need to deal with 2D arrays at some point
-        if isinstance(evaluated_child, np.ndarray) and len(evaluated_child.shape) == 1:
-            return self._averaging_function(evaluated_child)
-        # If not, no need to average
-        else:
-            return evaluated_child
-
-
 #
 # Methods to call Gradient and Divergence
 #
